requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
pandas==2.0.3
numpy==1.25.1
scikit-surprise==1.1.3